+++ nullstone.tf
@@ -25,6 +25,8 @@ resource "random_string" "resource_suffix" {
 
 locals {
   tags          = { for k, v in data.ns_workspace.this.tags : lower(k) => v }
+  stack_name    = data.ns_workspace.this.stack_name
+  env_name      = data.ns_workspace.this.env_name
   block_name    = data.ns_workspace.this.block_name
   block_ref     = data.ns_workspace.this.block_ref
   resource_name = "${local.block_ref}-${random_string.resource_suffix.result}"